Relocated controller files to correct folder name
